16-directional animation update
Added EntityAnimation to handle 16-directional animations: feed it a set
of 16 animations for an action and it'll automatically calculate what
animation to play given a direction vector and an action (animation
set). Examples on creating an animation set can be found in
PlayerAnimations.h. Implementation example can be found in
PlayerEntity.c

Updated player animation files accordingly.

Player's tracers now directly follow mouse with animation approximating
the angle instead of snapping to angle.